The datepicker functionality has been moved from the DateMask block to a separate calendarCard block.
CalendarCard block is an inline calendar in a separate block, which is not associated with any input. minDate and maxDate are not set in the CalendarCard block, but they can be set using the addMinMaxDates function. The CalendarCard block is set to select one date, but using the addDateRange function, you can set the choice of date ranges.
The DateMask block now includes only one addDateMask function, with which you can connect the date mask from another block.