Updating existing blocks.
The field, font Mixins, funcAddMultipleBlocks, and colorVariables blocks are combined into a baseVariablesAndMixins block.
In the baseVariablesAndMixins block, size variables is added.
The arrowForward and label blocks are removed.
The .pug file is deleted in the button block. Now the button block has no mixins.
Now all blocks with .pug mixins have the ability to add attributes (& attributes (attributes)).